% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C}
\alias{C}
\title{Single-cell sequencing of normal skin from human atopic dermatitis patients}
\format{
A data frame with 5991 genes and 472 cells:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE147424}
}
\usage{
C
}
\description{
Approximately 500 cells, randomly sub-sampled from the data.
}
\keyword{datasets}
