% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pDataC}
\alias{pDataC}
\title{Normal skin from human atopic dermatitis patients}
\format{
A data frame with 5991 genes and 472 cells:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE147424}
}
\usage{
pDataC
}
\description{
Labels corresponding to re-clustered cells in dataset `C`, following DESeq2-based differential marker-selection based on disease, given bulk/pseudobulk modality where bulk data are the normal and disease tissue samples from \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE65832}{Suarez-Farinas 2015 et. al.} and pseudobulk data are the normal and disease cells from \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE147424}{He 2020 et. al.}.  Differentially-expressed genes were selected based on adjusted p-value <= 1e-2.
}
\keyword{datasets}
