% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durian.R
\name{run_durian}
\alias{run_durian}
\title{Run DURIAN on input data}
\usage{
run_durian(
  path = NULL,
  scrabble_parameters = c(1, 1e-06, 1e-04),
  nIter_outer = 20,
  nIter_inner = 20,
  nSDCIters = 10000,
  emDiag = FALSE,
  nEM = 10,
  error_out_threshold = 1e-07,
  error_inner_threshold = 1e-05,
  C_fn = NULL,
  T_fn = NULL,
  pDataC_fn = NULL,
  scdata = NULL,
  bulkdata = NULL,
  metadata = NULL,
  seur = NULL,
  deconv_method = "dsLDA",
  MCNITER = 5000,
  MCNPARTITIONS = 5,
  MCNCHAINS = 2,
  MCBLOCKSIZE = 5,
  MCPLOT = TRUE,
  LDAPHILATENT = 2,
  LDABETAPSEUDO = 0,
  LDABETAEPS = 0,
  LDAALPHA = 0.1,
  LDANSCTHRESH = 3,
  SCRNSCTHRESH = 3,
  DECONVGENETHRESH = 0.01,
  SCRGENETHRESH = 0.3,
  MINCELLSTOPICCORP = 5,
  SCRSCALESC = "ident",
  SCRSCALEFACSC = 10000,
  LDASCALESC = "ident",
  LDASCALEBLK = "ident",
  LDASCALEFACSC = 1,
  LDASCALEFACBLK = 10000,
  LDAINITFLAVOR = "unif",
  LDARUNQC = FALSE,
  LDAVERBOSE = FALSE,
  MCTHINNING = 5,
  MCRMCHAINS = TRUE,
  MCBURNRATE = 0.25,
  deconvbenchmark = NULL,
  imputebenchmark = NULL,
  protectedgenes = c("dummygene"),
  maxT = 0,
  initscrabble = FALSE,
  allScr = TRUE,
  limrhat = 10,
  durianEps = 1e-06,
  saveDeconvolutionLog = FALSE,
  saveImputedStep = FALSE,
  saveImputationLog = TRUE,
  outerStats = FALSE,
  useIrlba = TRUE,
  runstability = FALSE
)
}
\arguments{
\item{path}{The path to the input data if not provided as an object.}

\item{scrabble_parameters}{The ADMM parameters}
}
\value{
The imputed data.
}
\description{
Run DURIAN on input data
}
