% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durian.R
\name{run_scrabble_m}
\alias{run_scrabble_m}
\title{Run multitask SCRABBLE (positive control)}
\usage{
run_scrabble_m(
  path,
  scrabble_parameters = c(1, 1e-06, 1e-04),
  nIter_outer = 20,
  nIter_inner = 20,
  nSDCIters = 10000,
  C_fn = "ldaC.csv",
  T_fn = "B.csv",
  pDataC_fn = "pDataC.csv",
  scdata = NULL,
  bulkdata = NULL,
  metadata = NULL,
  thetahat = NULL,
  imputebenchmark = NULL,
  scrcellids = NULL,
  scrgeneids = NULL,
  allScr = TRUE,
  error_out_threshold = 1e-07,
  error_inner_threshold = 1e-05,
  outerStats = FALSE,
  useIrlba = TRUE,
  runstability = FALSE
)
}
\arguments{
\item{path}{The path to the input data if not provided as an object.}

\item{scrabble_parameters}{The ADMM parameters}
}
\value{
The imputed data.
}
\description{
Run multitask SCRABBLE (positive control)
}
