% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durian.R
\name{subsetsc}
\alias{subsetsc}
\title{Get the cell and gene ids corresponding to gene expressed in at least `generate * ncells` cells and for those genes, the cells that have nonzero total expression}
\usage{
subsetsc(
  x = NULL,
  generate = 0.05,
  geneids = NULL,
  return_obj = FALSE,
  nsd = NULL
)
}
\description{
Get the cell and gene ids corresponding to gene expressed in at least `generate * ncells` cells and for those genes, the cells that have nonzero total expression
}
