% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T}
\alias{T}
\title{Bulk sequencing of normal skin from human atopic dermatitis patients}
\format{
A data frame with 6000 genes and 20 donors:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE65832}
}
\usage{
T
}
\description{
Normal skin tissue from 20 patients.
}
\keyword{datasets}
